New Project/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
uvicorn==0.24.0
fastapi==0.104.1
jinja2==3.1.2